README: Cleanup licensing (add CC BY-SA 4.0 for docs, etc.)
Before this commit, the README referenced a LICENSE.docs that didn't
exist locally.  This commit copies the LICENSE.code, LICENSE.docs, and
README blurb directly from the upstream repo.  The license text is
from master with updated copyright year in LICENSE.code.  The README
blurb is from my cleanup PR [1] with an updated copyright year and the
newlines removed.

[1]: https://github.com/docker/opensource/pull/98

Signed-off-by: W. Trevor King <wking@tremily.us>