WIP: circle.yml: Use Go 1.7
FIXME: The PATH echo will let me know how to setup
machine.environment.path, which needs to contain ~/go-1.7.4/bin.

Avoid [1]:

  go test -v -race ./...

  # github.com/docker/go-digest
  ./algorithm_test.go:43: t.Run undefined (type *testing.T has no field or method Run)
  ./verifiers_test.go:54: t.Run undefined (type *testing.T has no field or method Run)
  FAIL										github.com/docker/go-digest [build failed]

  go test -v -race ./... returned exit code 2

because *T.Run is new in Go 1.7 [2] and Circle is currently using
either Go 1.5.3 or 1.6.2 [3].  This commit installs Go 1.7 and caches
that installation for future Circle runs [4].  The test guard avoids
re-installing Go 1.7.4 when we already have it (due to Circle's
caching) [5].  The download URL and hash are from [6].

The GOROOT environment configuration [7] follows Go's recommendations
[8].

I've also included 'go version' calls for debugging (so it's easy to
know when we can drop this override).  The first call clears GOROOT so
we can see the system version.

[1]: https://circleci.com/gh/docker/go-digest/13#tests/containers/0
[2]: https://golang.org/doc/go1.7#testing
[3]: https://circleci.com/docs/language-go/#version
[4]: https://circleci.com/docs/configuration/#dependencies
[5]: https://circleci.com/docs/installing-custom-software/#avoid-recompiles
[6]: https://golang.org/dl/
[7]: https://circleci.com/docs/configuration/#environment
[8]: https://golang.org/doc/install#tarball_non_standard

Signed-off-by: W. Trevor King <wking@tremily.us>