digest: allow separators in algorithm field
In the past, we have had support for separators in the algorithm field
to allow parameterization of digest algorithms. A classic example is
`tarsum+sha256`. While this particular case is deprecated, support for
this case in the future must be allow in case we bring this back. This
will ensure that implementations these as valid digest, but correctly
report error when the algorithm is unsupported, rather than this being
treated as an invalid format.

We also widen the character set allowed in the hex encoded portion of
the digest to allow support for future encodings that are not hex-based.

Signed-off-by: Stephen J Day <stephen.day@docker.com>