algorithm: Deprecate ErrDigestInvalidLength
This error was added in 97212540 (Validate digest length on parsing,
2015-12-02) and motivated with [1]:

> The new error is mostly so that digest.Set doesn't need a special
> way to figure out if invalid digest had a algorithm prefix or not.

That doesn't make sense to me, because Digest.Validate returns
ErrDigestUnsupported when there is no algorithm prefix (I've added a
test to confirm this as well), and that seems orthogonal to encoding
validation (which is the only place ErrDigestInvalidLength could
matter).  From the other changes in [1], I think the issue was just
that invalid lengths weren't raising errors at all.

I see no reason to distinguish between invalid lengths, invalid
characters, and other invalid-encoded-portion errors, so this commit
collapses the error into ErrDigestInvalidFormat.

The code I'm removing assumed hex (with '.Size() * 2'), so this commit
removes that assumption. The sha256 and other hex-based algorithms are
still getting their lengths validated via anchoredEncodedRegexps.

[1]: https://github.com/docker/distribution/pull/1231

Signed-off-by: W. Trevor King <wking@tremily.us>